host-sp-comms: Factor out gimlet/cosmo shared code (#2203)
There are a number of identical helper functions repeated between the
`host-sp-comms` Gimlet and Cosmo BSP modules. These mostly deal with
reading IDs and VPD barcodes from EEPROMs. I found the duplicated code a
bit bothersome (and concerning) when changing `host-sp-comms`, since
changes have to be made in both Cosmo and Gimlet, and it's easy to make
one different from the other by mistake. Thus, I've moved this into a
shared module used on any compute sled.
